+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
 ## New features
 
+  1. add parallelisation to `GoF1DToys`
+  
 ## Backward incompatible
     
 ## Bug fixes